[IMPROVEMENT] Split CreateDiscussionView
